Module polars_lazy::dsl::functions

Functions

Function on multiple expressions.

Structs

DatetimeArgs

DurationArgs

Traits

Range

Functions

all

Selects all columns

all_exprs

Evaluate all the expressions with a bitwise and

any_exprs

Evaluate all the expressions with a bitwise or

apply_binary

argsort_by

Find the indexes that would sort these series in order of appearance. That means that the first Series will be used to determine the ordering until duplicates are found. Once duplicates are found, the next Series will be used and so on.

avg

Find the mean of all the values in this Expression.

cast

Cast expression.

coalesce

Folds the expressions from left to right keeping the first no null values.

col

Create a Column Expression based on a column name.

collect_all

Collect all LazyFrame computations.

cols

Select multiple columns by name

concat

Concat multiple

concat_lst

Concat lists entries.

cov

Compute the covariance between two columns.

date_range

Create a date range from a start and stop expression.

datetime

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

duration

fold_exprs

Accumulate over multiple columns horizontally / row wise.

is_not_null

IsNotNull expression.

is_null

IsNull expression

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

max

Find the maximum of all the values in this Expression.

max_exprs

Get the the maximum value per row

mean

Find the mean of all the values in this Expression.

median

Find the median of all the values in this Expression.

min

Find the minimum of all the values in this Expression.

min_exprs

not

Not expression.

pearson_corr

Compute the pearson correlation between two columns.

quantile

Find a specific quantile of all the values in this Expression.

range

Create a range literal.

reduce_exprs

repeat

Repeat a literal value n times.

sum

Sum all the values in this Expression.

sum_exprs

Get the the sum of the values per row