Struct polars_lazy::dsl::DatetimeArgs

pub struct DatetimeArgs {
    pub year: Expr,
    pub month: Expr,
    pub day: Expr,
    pub hour: Expr,
    pub minute: Expr,
    pub second: Expr,
    pub microsecond: Expr,
    pub time_unit: TimeUnit,
    pub time_zone: Option<String>,
    pub ambiguous: Expr,
}

Arguments used by datetime in order to produce an Expr of Datetime

Construct a DatetimeArgs with DatetimeArgs::new(y, m, d). This will set the other time units to lit(0). You can then set the other fields with the with_* methods, or use with_hms to set hour, minute, and second all at once.

Examples

use polars_plan::prelude::*;
// construct a DatetimeArgs set to July 20, 1969 at 20:17
let args = DatetimeArgs::new(lit(1969), lit(7), lit(20)).with_hms(lit(20), lit(17), lit(0));
// or
let args = DatetimeArgs::new(lit(1969), lit(7), lit(20)).with_hour(lit(20)).with_minute(lit(17));

// construct a DatetimeArgs using existing columns
let args = DatetimeArgs::new(lit(2023), col("month"), col("day"));

Fields
year: Exprmonth: Exprday: Exprhour: Exprminute: Exprsecond: Exprmicrosecond: Exprtime_unit: TimeUnittime_zone: Option<String>ambiguous: Expr
Implementations

impl DatetimeArgs

pub fn new(year: Expr, month: Expr, day: Expr) -> DatetimeArgs

Construct a new DatetimeArgs set to year, month, day

Other fields default to lit(0). Use the with_* methods to set them.

pub fn with_hms(self, hour: Expr, minute: Expr, second: Expr) -> DatetimeArgs

Set hour, minute, and second

Equivalent to

self.with_hour(hour)
    .with_minute(minute)
    .with_second(second)

pub fn with_year(self, n: Expr) -> DatetimeArgs

Set the year

pub fn with_month(self, n: Expr) -> DatetimeArgs

Set the month

pub fn with_day(self, n: Expr) -> DatetimeArgs

Set the day

pub fn with_hour(self, n: Expr) -> DatetimeArgs

Set the hour

pub fn with_minute(self, n: Expr) -> DatetimeArgs

Set the minute

pub fn with_second(self, n: Expr) -> DatetimeArgs

Set the second

pub fn with_microsecond(self, n: Expr) -> DatetimeArgs

Set the microsecond

pub fn with_time_unit(self, time_unit: TimeUnit) -> DatetimeArgs

pub fn with_time_zone(self, time_zone: Option<String>) -> DatetimeArgs

Available on crate feature timezones only.

pub fn with_ambiguous(self, ambiguous: Expr) -> DatetimeArgs

Available on crate feature timezones only.