polars_lazy::dsl

Struct ExprNameNameSpace

pub struct ExprNameNameSpace(/* private fields */);

Specialized expressions for modifying the name of existing expressions.

Implementations

impl ExprNameNameSpace

pub fn keep(self) -> Expr

Keep the original root name

fn example(df: LazyFrame) -> LazyFrame {
    df.select([
// even thought the alias yields a different column name,
// `keep` will make sure that the original column name is used
        col("*").alias("foo").name().keep()
])
}

pub fn map<F>(self, function: F) -> Expr
where F: Fn(&PlSmallStr) -> Result<PlSmallStr, PolarsError> + 'static + Send + Sync,

Define an alias by mapping a function over the original root column name.

pub fn map_udf(self, function: PythonObject) -> Expr

Define an alias by mapping a python lambda over the original root column name.

pub fn prefix(self, prefix: &str) -> Expr

Add a prefix to the root column name.

pub fn suffix(self, suffix: &str) -> Expr

Add a suffix to the root column name.

pub fn to_lowercase(self) -> Expr

Update the root column name to use lowercase characters.

pub fn to_uppercase(self) -> Expr

Update the root column name to use uppercase characters.

pub fn map_fields( self, function: Arc<dyn Fn(&str) -> PlSmallStr + Sync + Send>, ) -> Expr

pub fn map_fields_udf(self, function: PythonObject) -> Expr

pub fn prefix_fields(self, prefix: &str) -> Expr

pub fn suffix_fields(self, suffix: &str) -> Expr