polars_lazy::dsl
Struct ListNameSpace

pub struct ListNameSpace(pub Expr);

Specialized expressions for Series of DataType::List.

Tuple Fields
0: Expr
Implementations

impl ListNameSpace

pub fn any(self) -> Expr

Available on crate feature list_any_all only.

pub fn all(self) -> Expr

Available on crate feature list_any_all only.

pub fn drop_nulls(self) -> Expr

Available on crate feature list_drop_nulls only.

pub fn sample_n( self, n: Expr, with_replacement: bool, shuffle: bool, seed: Option<u64>, ) -> Expr

Available on crate feature list_sample only.

pub fn sample_fraction( self, fraction: Expr, with_replacement: bool, shuffle: bool, seed: Option<u64>, ) -> Expr

Available on crate feature list_sample only.

pub fn len(self) -> Expr

Return the number of elements in each list.

Null values are treated like regular elements in this context.

pub fn max(self) -> Expr

Compute the maximum of the items in every sublist.

pub fn min(self) -> Expr

Compute the minimum of the items in every sublist.

pub fn sum(self) -> Expr

Compute the sum the items in every sublist.

pub fn mean(self) -> Expr

Compute the mean of every sublist and return a Series of dtype Float64

pub fn median(self) -> Expr

pub fn std(self, ddof: u8) -> Expr

pub fn var(self, ddof: u8) -> Expr

pub fn sort(self, options: SortOptions) -> Expr

Sort every sublist.

pub fn reverse(self) -> Expr

Reverse every sublist

pub fn unique(self) -> Expr

Keep only the unique values in every sublist.

pub fn unique_stable(self) -> Expr

Keep only the unique values in every sublist.

pub fn n_unique(self) -> Expr

pub fn get(self, index: Expr, null_on_oob: bool) -> Expr

Get items in every sublist by index.

pub fn gather(self, index: Expr, null_on_oob: bool) -> Expr

Available on crate feature list_gather only.

Get items in every sublist by multiple indexes.

Arguments

null_on_oob: Return a null when an index is out of bounds. This behavior is more expensive than defaulting to returning an Error.

pub fn gather_every(self, n: Expr, offset: Expr) -> Expr

Available on crate feature list_gather only.

pub fn first(self) -> Expr

Get first item of every sublist.

pub fn last(self) -> Expr

Get last item of every sublist.

pub fn join(self, separator: Expr, ignore_nulls: bool) -> Expr

Join all string items in a sublist and place a separator between them.

Error

This errors if inner type of list != DataType::String.

pub fn arg_min(self) -> Expr

Return the index of the minimal value of every sublist

pub fn arg_max(self) -> Expr

Return the index of the maximum value of every sublist

pub fn diff(self, n: i64, null_behavior: NullBehavior) -> Expr

Available on crate feature diff only.

Diff every sublist.

pub fn shift(self, periods: Expr) -> Expr

Shift every sublist.

pub fn slice(self, offset: Expr, length: Expr) -> Expr

Slice every sublist.

pub fn head(self, n: Expr) -> Expr

Get the head of every sublist

pub fn tail(self, n: Expr) -> Expr

Get the tail of every sublist

pub fn to_array(self, width: usize) -> Expr

Available on crate feature dtype-array only.

Convert a List column into an Array column with the same inner data type.

pub fn to_struct( self, n_fields: ListToStructWidthStrategy, name_generator: Option<Arc<dyn Fn(usize) -> PlSmallStr + Send + Sync>>, upper_bound: usize, ) -> Expr

Available on crate feature list_to_struct only.

Convert this List to a Series of type Struct. The width will be determined according to ListToStructWidthStrategy and the names of the fields determined by the given name_generator.

Schema

A polars LazyFrame needs to know the schema at all time. The caller therefore must provide an upper_bound of struct fields that will be set. If this is incorrectly downstream operation may fail. For instance an all().sum() expression will look in the current schema to determine which columns to select.

pub fn contains<E>(self, other: E) -> Expr
where E: Into<Expr>,

Available on crate feature is_in only.

Check if the list array contain an element

pub fn count_matches<E>(self, element: E) -> Expr
where E: Into<Expr>,

Available on crate feature list_count only.

Count how often the value produced by element occurs.

pub fn union<E>(self, other: E) -> Expr
where E: Into<Expr>,

Available on crate feature list_sets only.

Return the SET UNION between both list arrays.

pub fn set_difference<E>(self, other: E) -> Expr
where E: Into<Expr>,

Available on crate feature list_sets only.

Return the SET DIFFERENCE between both list arrays.

pub fn set_intersection<E>(self, other: E) -> Expr
where E: Into<Expr>,

Available on crate feature list_sets only.

Return the SET INTERSECTION between both list arrays.

pub fn set_symmetric_difference<E>(self, other: E) -> Expr
where E: Into<Expr>,

Available on crate feature list_sets only.

Return the SET SYMMETRIC DIFFERENCE between both list arrays.