polars_lazy::dsl::udf

Function try_infer_udf_output_dtype

pub fn try_infer_udf_output_dtype(
    f: &dyn Fn(&[Column]) -> Result<Column, PolarsError>,
    input_fields: &[Field],
) -> Result<DataType, PolarsError>

Try to infer the output datatype of a UDF.

This will call the UDF in a few ways and see if it can get an output type without erroring.