Module polars_lazy::frame

Lazy variant of a DataFrame.

Structs

JoinOptions

LazyCsvReader

LazyFrame
Lazy abstraction over an eager DataFrame. It really is an abstraction over a logical plan. The methods of this struct will incrementally modify a logical plan until output is requested (via collect)

LazyGroupBy
Utility struct for lazy groupby operation.

OptState
State of the allowed optimizations

Traits

IntoLazy

Type Definitions

AllowedOptimizations
AllowedOptimizations