Struct polars_lazy::frame::LazyCsvReader

pub struct LazyCsvReader<'a> { /* fields omitted */ }

Implementations

impl<'a> LazyCsvReader<'a>

pub fn new(path: String) -> Self

pub fn with_skip_rows_after_header(self, offset: usize) -> Self

Skip this number of rows after the header location.

pub fn with_row_count(self, row_count: Option<RowCount>) -> Self

Add a row_count column.

pub fn with_n_rows(self, num_rows: Option<usize>) -> Self

Try to stop parsing when n rows are parsed. During multithreaded parsing the upper bound n cannot be guaranteed.

pub fn with_infer_schema_length(self, num_rows: Option<usize>) -> Self

Set the number of rows to use when inferring the csv schema. the default is 100 rows. Setting to None will do a full table scan, very slow.

pub fn with_ignore_parser_errors(self, ignore: bool) -> Self

Continue with next batch when a ParserError is encountered.

pub fn with_schema(self, schema: SchemaRef) -> Self

Set the CSV file’s schema

pub fn with_skip_rows(self, skip_rows: usize) -> Self

Skip the first n rows during parsing. The header will be parsed at row n.

pub fn with_dtype_overwrite(self, schema: Option<&'a Schema>) -> Self

Overwrite the schema with the dtypes in this given Schema. The given schema may be a subset of the total schema.

pub fn has_header(self, has_header: bool) -> Self

Set whether the CSV file has headers

pub fn with_delimiter(self, delimiter: u8) -> Self

Set the CSV file’s column delimiter as a byte character

pub fn with_comment_char(self, comment_char: Option<u8>) -> Self

Set the comment character. Lines starting with this character will be ignored.

pub fn with_quote_char(self, quote: Option<u8>) -> Self

Set the char used as quote char. The default is b'"'. If set to [None] quoting is disabled.

pub fn with_null_values(self, null_values: Option<NullValues>) -> Self

Set values that will be interpreted as missing/ null.

pub fn with_cache(self, cache: bool) -> Self

Cache the DataFrame after reading.

pub fn low_memory(self, toggle: bool) -> Self

Reduce memory usage in expensive of performance

pub fn with_rechunk(self, toggle: bool) -> Self

Rechunk the memory to contiguous chunks when parsing is done.

pub fn with_encoding(self, enc: CsvEncoding) -> Self

Set CsvEncoding

pub fn with_schema_modify<F>(self, f: F) -> Result<Self> where
F: Fn(Schema) -> Result<Schema>,

Modify a schema before we run the lazy scanning.

Important! Run this function latest in the builder!

pub fn finish_impl(self) -> Result<LazyFrame>

pub fn finish(self) -> Result<LazyFrame>