Trait polars_lazy::frame::LazyFileListReader

pub trait LazyFileListReader: Clone {
    // Required methods
    fn finish_no_glob(self) -> PolarsResult<LazyFrame>;
    fn path(&self) -> &Path;
    fn with_path(self, path: PathBuf) -> Self;
    fn rechunk(&self) -> bool;
    fn with_rechunk(self, toggle: bool) -> Self;
    fn n_rows(&self) -> Option<usize>;
    fn row_count(&self) -> Option<&RowCount>;

    // Provided methods
    fn finish(self) -> PolarsResult<LazyFrame> { ... }
    fn concat_impl(&self, lfs: Vec<LazyFrame>) -> PolarsResult<LazyFrame> { ... }
    fn cloud_options(&self) -> Option<&CloudOptions> { ... }
    fn glob(&self) -> PolarsResult<Option<GlobIterator>> { ... }
}

Reads LazyFrame from a filesystem or a cloud storage. Supports glob patterns.

Use LazyFileListReader::finish to get the final LazyFrame.

Required Methods

fn finish_no_glob(self) -> PolarsResult<LazyFrame>

Get the final LazyFrame. This method assumes, that path is not a glob.

It is recommended to always use LazyFileListReader::finish method.

fn path(&self) -> &Path

Path of the scanned file. It can be potentially a glob pattern.

fn with_path(self, path: PathBuf) -> Self

Set path of the scanned file. Support glob patterns.

fn rechunk(&self) -> bool

Rechunk the memory to contiguous chunks when parsing is done.

fn with_rechunk(self, toggle: bool) -> Self

Rechunk the memory to contiguous chunks when parsing is done.

fn n_rows(&self) -> Option<usize>

Try to stop parsing when n rows are parsed. During multithreaded parsing the upper bound n cannot be guaranteed.

fn row_count(&self) -> Option<&RowCount>

Add a row_count column.

Provided Methods

fn finish(self) -> PolarsResult<LazyFrame>

Get the final LazyFrame.

fn concat_impl(&self, lfs: Vec<LazyFrame>) -> PolarsResult<LazyFrame>

Recommended concatenation of LazyFrames from many input files.

This method should not take into consideration LazyFileListReader::n_rows nor LazyFileListReader::row_count.

fn cloud_options(&self) -> Option<&CloudOptions>

CloudOptions used to list files.

fn glob(&self) -> PolarsResult<Option<GlobIterator>>

Get list of files referenced by this reader.

Returns None if path is not a glob pattern.