Type Alias polars_lazy::frame::AllowedOptimizations

pub type AllowedOptimizations = OptState;

AllowedOptimizations

Aliased Type

struct AllowedOptimizations {

    pub projection_pushdown: bool,
    pub predicate_pushdown: bool,
    pub cluster_with_columns: bool,
    pub type_coercion: bool,
    pub simplify_expr: bool,
    pub file_caching: bool,
    pub slice_pushdown: bool,
    pub comm_subplan_elim: bool,
    pub comm_subexpr_elim: bool,
    pub streaming: bool,
    pub eager: bool,
    pub fast_projection: bool,
    pub row_estimate: bool,

}

Fields
projection_pushdown: bool

Only read columns that are used later in the query.

predicate_pushdown: bool

Apply predicates/filters as early as possible.

cluster_with_columns: bool

Cluster sequential with_columns calls to independent calls.

type_coercion: bool

Run many type coercion optimization rules until fixed point.

simplify_expr: bool

Run many expression optimization rules until fixed point.

file_caching: bool

Cache file reads.

slice_pushdown: bool

Pushdown slices/limits.

comm_subplan_elim: bool
Available on crate feature cse only.

Run common-subplan-elimination. This elides duplicate plans and caches their outputs.

comm_subexpr_elim: bool
Available on crate feature cse only.

Run common-subexpression-elimination. This elides duplicate expressions and caches their outputs.

streaming: bool

Run nodes that are capably of doing so on the streaming engine.

eager: bool

Run every node eagerly. This turns off multi-node optimizations.

fast_projection: bool

Replace simple projections with a faster inlined projection that skips the expression engine.

row_estimate: bool

Try to estimate the number of rows so that joins can determine which side to keep in memory.