Trait polars_lazy::physical_plan::executors::Executor

pub trait Executor: Send {
    fn execute(&mut self, cache: &mut ExecutionState) -> PolarsResult<DataFrame>;
}

Executors will evaluate physical expressions and collect them in a DataFrame.

Executors have other executors as input. By having a tree of executors we can execute the physical plan until the last executor is evaluated.

Required Methods

fn execute(&mut self, cache: &mut ExecutionState) -> PolarsResult<DataFrame>