Enum polars_lazy::prelude::LogicalPlan

pub enum LogicalPlan {

    Selection {
        input: Box<LogicalPlan, Global>,
        predicate: Expr,
    },
    Cache {
        input: Box<LogicalPlan, Global>,
        id: usize,
        count: usize,
    },
    Scan {
        paths: Arc<[PathBuf], Global>,
        file_info: FileInfo,
        predicate: Option<Expr>,
        file_options: FileScanOptions,
        scan_type: FileScan,
    },
    DataFrameScan {
        df: Arc<DataFrame, Global>,
        schema: Arc<Schema, Global>,
        output_schema: Option<Arc<Schema, Global>>,
        projection: Option<Arc<Vec<String, Global>, Global>>,
        selection: Option<Expr>,
    },
    Projection {
        expr: Vec<Expr, Global>,
        input: Box<LogicalPlan, Global>,
        schema: Arc<Schema, Global>,
        options: ProjectionOptions,
    },
    Aggregate {
        input: Box<LogicalPlan, Global>,
        keys: Arc<Vec<Expr, Global>, Global>,
        aggs: Vec<Expr, Global>,
        schema: Arc<Schema, Global>,
        apply: Option<Arc<dyn DataFrameUdf, Global>>,
        maintain_order: bool,
        options: Arc<GroupbyOptions, Global>,
    },
    Join {
        input_left: Box<LogicalPlan, Global>,
        input_right: Box<LogicalPlan, Global>,
        schema: Arc<Schema, Global>,
        left_on: Vec<Expr, Global>,
        right_on: Vec<Expr, Global>,
        options: Arc<JoinOptions, Global>,
    },
    HStack {
        input: Box<LogicalPlan, Global>,
        exprs: Vec<Expr, Global>,
        schema: Arc<Schema, Global>,
        options: ProjectionOptions,
    },
    Distinct {
        input: Box<LogicalPlan, Global>,
        options: DistinctOptions,
    },
    Sort {
        input: Box<LogicalPlan, Global>,
        by_column: Vec<Expr, Global>,
        args: SortArguments,
    },
    Slice {
        input: Box<LogicalPlan, Global>,
        offset: i64,
        len: u32,
    },
    MapFunction {
        input: Box<LogicalPlan, Global>,
        function: FunctionNode,
    },
    Union {
        inputs: Vec<LogicalPlan, Global>,
        options: UnionOptions,
    },
    Error {
        input: Box<LogicalPlan, Global>,
        err: ErrorStateSync,
    },
    ExtContext {
        input: Box<LogicalPlan, Global>,
        contexts: Vec<LogicalPlan, Global>,
        schema: Arc<Schema, Global>,
    },
    Sink {
        input: Box<LogicalPlan, Global>,
        payload: SinkType,
    },

}

Variants

Selection

Fields

input: Box<LogicalPlan, Global>

predicate: Expr

Filter on a boolean mask

Cache

Fields

input: Box<LogicalPlan, Global>

id: usize

count: usize

Cache the input at this point in the LP

Scan

Fields

paths: Arc<[PathBuf], Global>

file_info: FileInfo

predicate: Option<Expr>

file_options: FileScanOptions

scan_type: FileScan

DataFrameScan

Fields

df: Arc<DataFrame, Global>

schema: Arc<Schema, Global>

output_schema: Option<Arc<Schema, Global>>

projection: Option<Arc<Vec<String, Global>, Global>>

selection: Option<Expr>

In memory DataFrame

Projection

Fields

expr: Vec<Expr, Global>

input: Box<LogicalPlan, Global>

schema: Arc<Schema, Global>

options: ProjectionOptions

Column selection

Aggregate

Fields

input: Box<LogicalPlan, Global>

keys: Arc<Vec<Expr, Global>, Global>

aggs: Vec<Expr, Global>

schema: Arc<Schema, Global>

apply: Option<Arc<dyn DataFrameUdf, Global>>

maintain_order: bool

options: Arc<GroupbyOptions, Global>

Groupby aggregation

Join

Fields

input_left: Box<LogicalPlan, Global>

input_right: Box<LogicalPlan, Global>

schema: Arc<Schema, Global>

left_on: Vec<Expr, Global>

right_on: Vec<Expr, Global>

options: Arc<JoinOptions, Global>

Join operation

HStack

Fields

input: Box<LogicalPlan, Global>

exprs: Vec<Expr, Global>

schema: Arc<Schema, Global>

options: ProjectionOptions

Adding columns to the table without a Join

Distinct

Fields

input: Box<LogicalPlan, Global>

options: DistinctOptions

Remove duplicates from the table

Sort

Fields

input: Box<LogicalPlan, Global>

by_column: Vec<Expr, Global>

args: SortArguments

Sort the table

Slice

Fields

input: Box<LogicalPlan, Global>

offset: i64

len: u32

Slice the table

MapFunction

Fields

input: Box<LogicalPlan, Global>

function: FunctionNode

A (User Defined) Function

Union

Fields

inputs: Vec<LogicalPlan, Global>

options: UnionOptions

Error

Fields

input: Box<LogicalPlan, Global>

err: ErrorStateSync

Catches errors and throws them later

ExtContext

Fields

input: Box<LogicalPlan, Global>

contexts: Vec<LogicalPlan, Global>

schema: Arc<Schema, Global>

This allows expressions to access other tables

Sink

Fields

input: Box<LogicalPlan, Global>

payload: SinkType

Implementations

impl LogicalPlan

pub fn dot( &self, acc_str: &mut String, id: (usize, usize), prev_node: DotNode<'_>, id_map: &mut HashMap<String, String, RandomState, Global> ) -> Result<(), Error>

Arguments

id - (branch, id) Used to make sure that the dot boxes are distinct. branch is an id per join/union branch id is incremented by the depth traversal of the tree.

impl LogicalPlan

pub fn schema(&self) -> Result<Cow<'_, Arc<Schema, Global>>, PolarsError>

impl LogicalPlan

pub fn describe(&self) -> String

pub fn to_alp( self ) -> Result<(Node, Arena<ALogicalPlan>, Arena<AExpr>), PolarsError>