Module polars_lazy::dsl

Domain specific language for the Lazy api.

Re-exports

pub use crate::frame::IntoLazy;

pub use crate::functions::*;

Modules

string

Structs

NoEq

Wrapper type that indicates that the inner type is not equal to anything

StrpTimeOptions

When

Intermediate state of when(..).then(..).otherwise(..) expr.

WhenThen

Intermediate state of when(..).then(..).otherwise(..) expr.

WhenThenThen

Intermediate state of chain when then exprs.

Enums

AggExpr

Excluded

Expr

Queries consists of multiple expressions.

Operator

Traits

BinaryUdfOutputField

FunctionOutputField

SeriesBinaryUdf

A wrapper trait for any binary closure Fn(Series, Series) -> Result<Series>

SeriesUdf

A wrapper trait for any closure Fn(Vec<Series>) -> Result<Series>

Functions

apply_multiple

Apply a function/closure over the groups of multiple columns. This should only be used in a groupby aggregation.

binary_expr

count

Count expression

first

First column in DataFrame

last

Last column in DataFrame

lit

Create a Literal Expression from L

map_list_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

map_multiple

Apply a function/closure over multiple columns once the logical plan get executed.

ternary_expr

when

Start a when-then-otherwise expression

Type Definitions

GetOutput