polars_lazy::dsl

Module functions

Functions

Function on multiple expressions.

Structs

DatetimeArgs

Arguments used by datetime in order to produce an Expr of Datetime

DurationArgs

Arguments used by duration in order to produce an Expr of Duration

Functions

all

Selects all columns. Shorthand for col("*").

all_horizontal

Create a new column with the bitwise-and of the elements in each row.

any_horizontal

Create a new column with the bitwise-or of the elements in each row.

apply_binary

Like map_binary, but used in a group_by-aggregation context.

avg

Find the mean of all the values in the column named name. Alias for mean.

cast

Casts the column given by Expr to a different type.

coalesce

Folds the expressions from left to right keeping the first non-null values.

col

Create a Column Expression based on a column name.

collect_all

Collect all LazyFrame computations.

cols

Select multiple columns by name.

concat

Concat multiple LazyFrames vertically.

concat_arr

Horizontally concatenate columns into a single array-type column.

concat_expr

concat_lf_horizontal

Concat LazyFrames horizontally.

concat_list

Concat lists entries.

datetime

Construct a column of Datetime from the provided DatetimeArgs.

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

duration

Construct a column of Duration from the provided DurationArgs

fold_exprs

Accumulate over multiple columns horizontally / row wise.

index_cols

Select multiple columns by index.

is_not_null

A column which is false wherever expr is null, true elsewhere.

is_null

A column which is true wherever expr is null, false elsewhere.

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

max

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

mean

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

median

Find the median of all the values in the column named name. Shorthand for col(name).median().

min

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

not

Negates a boolean column.

quantile

Find a specific quantile of all the values in the column named name.

reduce_exprs

Analogous to Iterator::reduce.

repeat

Create a column of length n containing n copies of the literal value.

sum

Sum all the values in the column named name. Shorthand for col(name).sum().