polars_lazy::dsl
Module functions

Functions

Function on multiple expressions.

Structs

DatetimeArgs

Arguments used by datetime in order to produce an Expr of Datetime

DurationArgs

Arguments used by duration in order to produce an Expr of Duration

Functions

all

Selects all columns. Shorthand for col("*").

all_horizontal

Create a new column with the bitwise-and of the elements in each row.

any_horizontal

Create a new column with the bitwise-or of the elements in each row.

apply_binary

Like map_binary, but used in a group_by-aggregation context.

arange

Generate a range of integers.

arg_sort_byrange

Find the indexes that would sort these series in order of appearance.

arg_wherearg_where

Get the indices where condition evaluates true.

as_struct

Take several expressions and collect them into a StructChunked.

avg

Find the mean of all the values in the column named name. Alias for mean.

cast

Casts the column given by Expr to a different type.

coalesce

Folds the expressions from left to right keeping the first non-null values.

col

Create a Column Expression based on a column name.

collect_all

Collect all LazyFrame computations.

cols

Select multiple columns by name.

concat

Concat multiple LazyFrames vertically.

concat_expr

concat_lf_diagonaldiagonal_concat

Concat LazyFrames diagonally. Calls concat internally.

concat_lf_horizontal

Concat LazyFrames horizontally.

concat_list

Concat lists entries.

concat_strconcat_str and strings

Horizontally concat string columns in linear time

cov

Compute the covariance between two columns.

cum_fold_exprsdtype-struct

Accumulate over multiple columns horizontally / row wise.

cum_reduce_exprsdtype-struct

Accumulate over multiple columns horizontally / row wise.

date_rangetemporal

Create a date range from a start and stop expression.

date_rangestemporal

Create a column of date ranges from a start and stop expression.

datetime

Construct a column of Datetime from the provided DatetimeArgs.

datetime_rangedtype-datetime

Create a datetime range from a start and stop expression.

datetime_rangesdtype-datetime

Create a column of datetime ranges from a start and stop expression.

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

duration

Construct a column of Duration from the provided DurationArgs

fold_exprs

Accumulate over multiple columns horizontally / row wise.

format_strconcat_str and strings

Format the results of an array of expressions using a format string

index_cols

Select multiple columns by index.

int_range

Generate a range of integers.

int_ranges

Generate a range of integers for each row of the input columns.

is_not_null

A column which is false wherever expr is null, true elsewhere.

is_null

A column which is true wherever expr is null, false elsewhere.

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

max

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

mean

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

median

Find the median of all the values in the column named name. Shorthand for col(name).median().

min

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

not

Negates a boolean column.

pearson_corr

Compute the pearson correlation between two columns.

quantile

Find a specific quantile of all the values in the column named name.

reduce_exprs

Analogous to Iterator::reduce.

repeat

Create a column of length n containing n copies of the literal value.

rolling_corrrolling_window

rolling_covrolling_window

spearman_rank_corrrank and propagate_nans

Compute the spearman rank correlation between two columns. Missing data will be excluded from the computation.

sum

Sum all the values in the column named name. Shorthand for col(name).sum().

time_rangedtype-time

Generate a time range.

time_rangesdtype-time

Create a column of time ranges from a start and stop expression.