Module polars_lazy::dsl::functions

Functions

Function on multiple expressions.

Structs

DatetimeArgs

DurationArgs

Traits

Range

Functions

all

Selects all columns. Shorthand for col("*").

all_exprs

Create a new column with the the bitwise-and of the elements in each row.

any_exprs

Create a new column with the the bitwise-or of the elements in each row.

apply_binary

Like map_binary, but used in a groupby-aggregation context.

arangearange

Create list entries that are range arrays

arg_sort_by

Find the indexes that would sort these series in order of appearance. That means that the first Series will be used to determine the ordering until duplicates are found. Once duplicates are found, the next Series will be used and so on.

arg_wherearg_where

Get the indices where condition evaluates true.

as_structdtype-struct

Take several expressions and collect them into a StructChunked.

avg

Find the mean of all the values in the column named name. Alias for mean.

cast

Casts the column given by Expr to a different type.

coalesce

Folds the expressions from left to right keeping the first non-null values.

col

Create a Column Expression based on a column name.

collect_all

Collect all LazyFrame computations.

cols

Select multiple columns by name.

concat

Concat multiple

concat_lst

Concat lists entries.

concat_strconcat_str and strings

Horizontally concat string columns in linear time

cov

Compute the covariance between two columns.

cumfold_exprsdtype-struct

Accumulate over multiple columns horizontally / row wise.

cumreduce_exprsdtype-struct

Accumulate over multiple columns horizontally / row wise.

date_rangetemporal

Create a date range, named name, from a start and stop expression.

datetimetemporal

Construct a column of Datetime from the provided args.

diag_concat_lfdiagonal_concat

Concat LazyFrames diagonally. Calls [concat] internally.

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

durationtemporal

Construct a column of Duration from the provided args.

fold_exprs

Accumulate over multiple columns horizontally / row wise.

format_strconcat_str and strings

Format the results of an array of expressions using a format string

is_not_null

A column which is false wherever expr is null, true elsewhere.

is_null

A column which is true wherever expr is null, false elsewhere.

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

max

Find the maximum of all the values in the column named name. Shorthand for col(name).max().

max_exprs

Create a new column with the the maximum value per row.

mean

Find the mean of all the values in the column named name. Shorthand for col(name).mean().

median

Find the median of all the values in the column named name. Shorthand for col(name).median().

min

Find the minimum of all the values in the column named name. Shorthand for col(name).min().

min_exprs

Create a new column with the the minimum value per row.

not

Negates a boolean column.

pearson_corr

Compute the pearson correlation between two columns.

quantile

Find a specific quantile of all the values in the column named name.

range

Create a range literal.

reduce_exprs

Analogous to Iterator::reduce.

repeat

Create a column of length n containing n copies of the literal value. Generally you won’t need this function, as lit(value) already represents a column containing only value whose length is automatically set to the correct number of rows.

spearman_rank_corrrank and propagate_nans

Compute the spearman rank correlation between two columns. Missing data will be excluded from the computation.

sum

Sum all the values in the column named name. Shorthand for col(name).sum().

sum_exprs

Create a new column with the the sum of the values in each row.