Module polars_lazy::dsl::functions

Functions

Function on multiple expressions.

Structs

DatetimeArgs

DurationArgs

Traits

Range

Functions

all

Selects all columns

all_exprs

Evaluate all the expressions with a bitwise and

any_exprs

Evaluate all the expressions with a bitwise or

apply_binary

arangearange

Create list entries that are range arrays

arg_sort_by

Find the indexes that would sort these series in order of appearance. That means that the first Series will be used to determine the ordering until duplicates are found. Once duplicates are found, the next Series will be used and so on.

arg_wherearg_where

Get the indices where condition evaluates true.

as_structdtype-struct

Take several expressions and collect them into a StructChunked.

avg

Find the mean of all the values in this Expression.

cast

Cast expression.

coalesce

Folds the expressions from left to right keeping the first no null values.

col

Create a Column Expression based on a column name.

collect_all

Collect all LazyFrame computations.

cols

Select multiple columns by name

concat

Concat multiple

concat_lst

Concat lists entries.

concat_strconcat_str and strings

Horizontally concat string columns in linear time

cov

Compute the covariance between two columns.

cumfold_exprsdtype-struct

Accumulate over multiple columns horizontally / row wise.

cumreduce_exprsdtype-struct

Accumulate over multiple columns horizontally / row wise.

date_rangetemporal

Create a date range from a start and stop expression.

datetimetemporal

diag_concat_lfdiagonal_concat

Concat LazyFrames diagonally. Calls [concat] internally.

dtype_col

Select multiple columns by dtype.

dtype_cols

Select multiple columns by dtype.

durationtemporal

fold_exprs

Accumulate over multiple columns horizontally / row wise.

format_strconcat_str and strings

Format the results of an array of expressions using a format string

is_not_null

IsNotNull expression.

is_null

IsNull expression

map_binary

Apply a closure on the two columns that are evaluated from Expr a and Expr b.

max

Find the maximum of all the values in this Expression.

max_exprs

Get the the maximum value per row

mean

Find the mean of all the values in this Expression.

median

Find the median of all the values in this Expression.

min

Find the minimum of all the values in this Expression.

min_exprs

not

Not expression.

pearson_corr

Compute the pearson correlation between two columns.

quantile

Find a specific quantile of all the values in this Expression.

range

Create a range literal.

reduce_exprs

repeat

Repeat a literal value n times.

spearman_rank_corrrank and propagate_nans

Compute the spearman rank correlation between two columns. Missing data will be excluded from the computation.

sum

Sum all the values in this Expression.

sum_exprs

Get the the sum of the values per row