Struct polars_lazy::dsl::ArrayNameSpace

pub struct ArrayNameSpace(pub Expr);

Specialized expressions for Series of DataType::Array.

Tuple Fields
0: Expr
Implementations

impl ArrayNameSpace

pub fn max(self) -> Expr

Compute the maximum of the items in every subarray.

pub fn min(self) -> Expr

Compute the minimum of the items in every subarray.

pub fn sum(self) -> Expr

Compute the sum of the items in every subarray.

pub fn std(self, ddof: u8) -> Expr

Compute the std of the items in every subarray.

pub fn var(self, ddof: u8) -> Expr

Compute the var of the items in every subarray.

pub fn median(self) -> Expr

Compute the median of the items in every subarray.

pub fn unique(self) -> Expr

Keep only the unique values in every sub-array.

pub fn unique_stable(self) -> Expr

Keep only the unique values in every sub-array.

pub fn n_unique(self) -> Expr

pub fn to_list(self) -> Expr

Cast the Array column to List column with the same inner data type.

pub fn all(self) -> Expr

Available on crate feature array_any_all only.

Evaluate whether all boolean values are true for every subarray.

pub fn any(self) -> Expr

Available on crate feature array_any_all only.

Evaluate whether any boolean value is true for every subarray

pub fn sort(self, options: SortOptions) -> Expr

pub fn reverse(self) -> Expr

pub fn arg_min(self) -> Expr

pub fn arg_max(self) -> Expr

pub fn get(self, index: Expr, null_on_oob: bool) -> Expr

Get items in every sub-array by index.

pub fn join(self, separator: Expr, ignore_nulls: bool) -> Expr

Join all string items in a sub-array and place a separator between them.

Error

Raise if inner type of array is not DataType::String.

pub fn contains<E>(self, other: E) -> Expr
where E: Into<Expr>,

Available on crate feature is_in only.

Check if the sub-array contains specific element

pub fn count_matches<E>(self, element: E) -> Expr
where E: Into<Expr>,

Available on crate feature array_count only.

Count how often the value produced by element occurs.

pub fn to_struct( self, name_generator: Option<Arc<dyn Fn(usize) -> SmartString<LazyCompact> + Send + Sync>> ) -> Expr

Available on crate feature array_to_struct only.

pub fn shift(self, n: Expr) -> Expr

Shift every sub-array.