polars_lazy::dsl

Struct MetaNameSpace

pub struct MetaNameSpace(/* private fields */);

Specialized expressions for Categorical dtypes.

Implementations

impl MetaNameSpace

pub fn pop( self, schema: Option<&Schema<DataType>>, ) -> Result<Vec<Expr>, PolarsError>

Pop latest expression and return the input(s) of the popped expression.

pub fn root_names(&self) -> Vec<PlSmallStr>

Get the root column names.

pub fn is_simple_projection(&self, schema: Option<&Schema<DataType>>) -> bool

A projection that only takes a column or a column + alias.

pub fn output_name(&self) -> Result<PlSmallStr, PolarsError>

Get the output name of this expression.

pub fn undo_aliases(self) -> Expr

Undo any renaming operation like alias, keep_name.

pub fn has_multiple_outputs(&self) -> bool

Indicate if this expression expands to multiple expressions.

pub fn is_column(&self) -> bool

Indicate if this expression is a basic (non-regex) column.

pub fn is_column_selection(&self, allow_aliasing: bool) -> bool

Indicate if this expression only selects columns; the presence of any transform operations will cause the check to return false, though aliasing of the selected columns is optionally allowed.

pub fn is_literal(&self, allow_aliasing: bool) -> bool

Indicate if this expression represents a literal value (optionally aliased).

pub fn is_regex_projection(&self) -> bool

Indicate if this expression expands to multiple expressions with regex expansion.

pub fn into_tree_formatter( self, display_as_dot: bool, schema: Option<&Schema<DataType>>, ) -> Result<impl Display, PolarsError>

Get a hold to an implementor of the Display trait that will format as the expression as a tree