Module polars_lazy::frame

Lazy variant of a DataFrame.

Modules

pivotpivot

Polars lazy does not implement a pivot because it is impossible to know the schema without materializing the whole dataset. This makes a pivot quite a terrible operation for performant workflows. An optimization can never be pushed down passed a pivot.

Structs

InProcessQueryNon-WebAssembly

JoinBuilder

LazyCsvReadercsv

LazyFrame

Lazy abstraction over an eager DataFrame. It really is an abstraction over a logical plan. The methods of this struct will incrementally modify a logical plan until output is requested (via collect).

LazyGroupBy

Utility struct for lazy group_by operation.

LazyJsonLineReaderjson

OptState

State of the allowed optimizations

ScanArgsAnonymous

ScanArgsIpcipc

ScanArgsParquetparquet

Traits

IntoLazy

LazyFileListReader

Reads LazyFrame from a filesystem or a cloud storage. Supports glob patterns.

Type Aliases

AllowedOptimizations

AllowedOptimizations

PathIterator