Struct polars_lazy::frame::LazyCsvReader

pub struct LazyCsvReader<'a> { /* private fields */ }

Available on crate feature csv only.

Implementations

impl<'a> LazyCsvReader<'a>

pub fn new_paths(paths: Arc<[PathBuf]>) -> Self

pub fn new(path: impl AsRef<Path>) -> Self

pub fn with_skip_rows_after_header(self, offset: usize) -> Self

Skip this number of rows after the header location.

pub fn with_row_index(self, row_index: Option<RowIndex>) -> Self

Add a row index column.

pub fn with_n_rows(self, num_rows: Option<usize>) -> Self

Try to stop parsing when n rows are parsed. During multithreaded parsing the upper bound n cannot be guaranteed.

pub fn with_infer_schema_length(self, num_rows: Option<usize>) -> Self

Set the number of rows to use when inferring the csv schema. the default is 100 rows. Setting to None will do a full table scan, very slow.

pub fn with_ignore_errors(self, ignore: bool) -> Self

Continue with next batch when a ParserError is encountered.

pub fn with_schema(self, schema: Option<SchemaRef>) -> Self

Set the CSV file’s schema

pub fn with_skip_rows(self, skip_rows: usize) -> Self

Skip the first n rows during parsing. The header will be parsed at row n.

pub fn with_dtype_overwrite(self, schema: Option<&'a Schema>) -> Self

Overwrite the schema with the dtypes in this given Schema. The given schema may be a subset of the total schema.

pub fn has_header(self, has_header: bool) -> Self

Set whether the CSV file has headers

pub fn with_separator(self, separator: u8) -> Self

Set the CSV file’s column separator as a byte character

pub fn with_comment_prefix(self, comment_prefix: Option<&str>) -> Self

Set the comment prefix for this instance. Lines starting with this prefix will be ignored.

pub fn with_quote_char(self, quote: Option<u8>) -> Self

Set the char used as quote char. The default is b'"'. If set to [None] quoting is disabled.

pub fn with_end_of_line_char(self, eol_char: u8) -> Self

Set the char used as end of line. The default is b'\n'.

pub fn with_null_values(self, null_values: Option<NullValues>) -> Self

Set values that will be interpreted as missing/ null.

pub fn with_missing_is_null(self, missing_is_null: bool) -> Self

Treat missing fields as null.

pub fn with_cache(self, cache: bool) -> Self

Cache the DataFrame after reading.

pub fn low_memory(self, toggle: bool) -> Self

Reduce memory usage at the expense of performance

pub fn with_encoding(self, enc: CsvEncoding) -> Self

Set CsvEncoding

pub fn with_try_parse_dates(self, toggle: bool) -> Self

Available on crate feature temporal only.

Automatically try to parse dates/datetimes and time. If parsing fails, columns remain of dtype [DataType::String].

pub fn raise_if_empty(self, toggle: bool) -> Self

Raise an error if CSV is empty (otherwise return an empty frame)

pub fn truncate_ragged_lines(self, toggle: bool) -> Self

Truncate lines that are longer than the schema.

pub fn with_schema_modify<F>(self, f: F) -> PolarsResult<Self>
where F: Fn(Schema) -> PolarsResult<Schema>,

Modify a schema before we run the lazy scanning.

Important! Run this function latest in the builder!