Struct polars_lazy::frame::LazyJsonLineReader

pub struct LazyJsonLineReader { /* private fields */ }

Available on crate feature json only.

Implementations

impl LazyJsonLineReader

pub fn new_paths(paths: Arc<[PathBuf]>) -> Self

pub fn new(path: impl AsRef<Path>) -> Self

pub fn with_row_index(self, row_index: Option<RowIndex>) -> Self

Add a row index column.

pub fn with_ignore_errors(self, ignore_errors: bool) -> Self

Set values as Null if parsing fails because of schema mismatches.

pub fn with_n_rows(self, num_rows: Option<usize>) -> Self

Try to stop parsing when n rows are parsed. During multithreaded parsing the upper bound n cannot be guaranteed.

pub fn with_infer_schema_length(self, num_rows: Option<usize>) -> Self

Set the number of rows to use when inferring the json schema. the default is 100 rows. Ignored when the schema is specified explicitly using Self::with_schema. Setting to None will do a full table scan, very slow.

pub fn with_schema(self, schema: Option<SchemaRef>) -> Self

Set the JSON file’s schema

pub fn low_memory(self, toggle: bool) -> Self

Reduce memory usage at the expense of performance

pub fn with_batch_size(self, batch_size: Option<NonZeroUsize>) -> Self