Trait polars_lazy::physical_plan::expressions::PhysicalExpr

pub trait PhysicalExpr: Send + Sync {
    fn evaluate(
        &self, 
        df: &DataFrame, 
        _state: &ExecutionState
    ) -> Result<Series>;

    fn to_field(&self, input_schema: &Schema) -> Result<Field>;

    fn as_expression(&self) -> &Expr { ... }

    fn evaluate_on_groups<'a>(
        &self, 
        df: &DataFrame, 
        groups: &'a GroupTuples, 
        state: &ExecutionState
    ) -> Result<(Series, Cow<'a, GroupTuples>)> { ... }

    fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation> { ... }
}

Take a DataFrame and evaluate the expressions. Implement this for Column, lt, eq, etc

Required methods

fn evaluate(&self, df: &DataFrame, _state: &ExecutionState) -> Result<Series>

Take a DataFrame and evaluate the expression.

fn to_field(&self, input_schema: &Schema) -> Result<Field>

Get the output field of this expr

Loading content...
Provided methods

fn as_expression(&self) -> &Expr

fn evaluate_on_groups<'a>(
&self,
df: &DataFrame,
groups: &'a GroupTuples,
state: &ExecutionState
) -> Result<(Series, Cow<'a, GroupTuples>)>

Some expression that are not aggregations can be done per group Think of sort, slice, etc.

defaults to ignoring the group

fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation>

Convert to a aggregation expression. This can only be done for the final expressions that produce an aggregated result.

The expression sum, min, max etc can be called as evaluate in the standard context, or during a groupby execution, this method is called to convert them to an AggPhysicalExpr

Loading content...