Enum polars_lazy::prelude::LogicalPlan

pub enum LogicalPlan {

    PythonScan {
        options: PythonOptions,
    },
    Selection {
        input: Box<LogicalPlan>,
        predicate: Expr,
    },
    Cache {
        input: Box<LogicalPlan>,
        id: usize,
        cache_hits: u32,
    },
    Scan {
        paths: Arc<[PathBuf]>,
        file_info: FileInfo,
        predicate: Option<Expr>,
        file_options: FileScanOptions,
        scan_type: FileScan,
    },
    DataFrameScan {
        df: Arc<DataFrame>,
        schema: Arc<Schema>,
        output_schema: Option<Arc<Schema>>,
        projection: Option<Arc<Vec<String>>>,
        selection: Option<Expr>,
    },
    Projection {
        expr: Vec<Expr>,
        input: Box<LogicalPlan>,
        schema: Arc<Schema>,
        options: ProjectionOptions,
    },
    Aggregate {
        input: Box<LogicalPlan>,
        keys: Arc<Vec<Expr>>,
        aggs: Vec<Expr>,
        schema: Arc<Schema>,
        apply: Option<Arc<dyn DataFrameUdf>>,
        maintain_order: bool,
        options: Arc<GroupbyOptions>,
    },
    Join {
        input_left: Box<LogicalPlan>,
        input_right: Box<LogicalPlan>,
        schema: Arc<Schema>,
        left_on: Vec<Expr>,
        right_on: Vec<Expr>,
        options: Arc<JoinOptions>,
    },
    HStack {
        input: Box<LogicalPlan>,
        exprs: Vec<Expr>,
        schema: Arc<Schema>,
        options: ProjectionOptions,
    },
    Distinct {
        input: Box<LogicalPlan>,
        options: DistinctOptions,
    },
    Sort {
        input: Box<LogicalPlan>,
        by_column: Vec<Expr>,
        args: SortArguments,
    },
    Slice {
        input: Box<LogicalPlan>,
        offset: i64,
        len: u64,
    },
    MapFunction {
        input: Box<LogicalPlan>,
        function: FunctionNode,
    },
    Union {
        inputs: Vec<LogicalPlan>,
        options: UnionOptions,
    },
    HConcat {
        inputs: Vec<LogicalPlan>,
        schema: Arc<Schema>,
        options: HConcatOptions,
    },
    Error {
        input: Box<LogicalPlan>,
        err: ErrorState,
    },
    ExtContext {
        input: Box<LogicalPlan>,
        contexts: Vec<LogicalPlan>,
        schema: Arc<Schema>,
    },
    Sink {
        input: Box<LogicalPlan>,
        payload: SinkType,
    },

}

Variants

PythonScan

Available on crate feature python only.

Fields

options: PythonOptions

Selection

Filter on a boolean mask

Fields

input: Box<LogicalPlan>

predicate: Expr

Cache

Cache the input at this point in the LP

Fields

input: Box<LogicalPlan>

id: usize

cache_hits: u32

Scan

Fields

paths: Arc<[PathBuf]>

file_info: FileInfo

predicate: Option<Expr>

file_options: FileScanOptions

scan_type: FileScan

DataFrameScan

In memory DataFrame

Fields

df: Arc<DataFrame>

schema: Arc<Schema>

output_schema: Option<Arc<Schema>>

projection: Option<Arc<Vec<String>>>

selection: Option<Expr>

Projection

Column selection

Fields

expr: Vec<Expr>

input: Box<LogicalPlan>

schema: Arc<Schema>

options: ProjectionOptions

Aggregate

Groupby aggregation

Fields

input: Box<LogicalPlan>

keys: Arc<Vec<Expr>>

aggs: Vec<Expr>

schema: Arc<Schema>

apply: Option<Arc<dyn DataFrameUdf>>

maintain_order: bool

options: Arc<GroupbyOptions>

Join

Join operation

Fields

input_left: Box<LogicalPlan>

input_right: Box<LogicalPlan>

schema: Arc<Schema>

left_on: Vec<Expr>

right_on: Vec<Expr>

options: Arc<JoinOptions>

HStack

Adding columns to the table without a Join

Fields

input: Box<LogicalPlan>

exprs: Vec<Expr>

schema: Arc<Schema>

options: ProjectionOptions

Distinct

Remove duplicates from the table

Fields

input: Box<LogicalPlan>

options: DistinctOptions

Sort

Sort the table

Fields

input: Box<LogicalPlan>

by_column: Vec<Expr>

args: SortArguments

Slice

Slice the table

Fields

input: Box<LogicalPlan>

offset: i64

len: u64

MapFunction

A (User Defined) Function

Fields

input: Box<LogicalPlan>

function: FunctionNode

Union

Fields

inputs: Vec<LogicalPlan>

options: UnionOptions

HConcat

Horizontal concatenation of multiple plans

Fields

inputs: Vec<LogicalPlan>

schema: Arc<Schema>

options: HConcatOptions

Error

Catches errors and throws them later

Fields

input: Box<LogicalPlan>

err: ErrorState

ExtContext

This allows expressions to access other tables

Fields

input: Box<LogicalPlan>

contexts: Vec<LogicalPlan>

schema: Arc<Schema>

Sink

Fields

input: Box<LogicalPlan>

payload: SinkType

Implementations

impl LogicalPlan

pub fn dot( &self, acc_str: &mut String, id: (usize, usize), prev_node: DotNode<'_>, id_map: &mut HashMap<String, String, RandomState> ) -> Result<(), Error>

Arguments

id - (branch, id) Used to make sure that the dot boxes are distinct. branch is an id per join/union branch id is incremented by the depth traversal of the tree.

impl LogicalPlan

pub fn schema(&self) -> Result<Cow<'_, Arc<Schema>>, PolarsError>

impl LogicalPlan

pub fn describe(&self) -> String

pub fn describe_tree_format(&self) -> String

pub fn to_alp( self ) -> Result<(Node, Arena<ALogicalPlan>, Arena<AExpr>), PolarsError>